Feature: add hash of original file when converting
The hash is used for checking whether the original file has changed
since conversion. If identical, conversion can be skipped. If different,
we error out with a message.

Note that generating the portable library is a non-destructive operation
-- after the non-conversion simple copies are done anyway. Because of
this, it's possible to have "extra" files in the portable library if you
regenerate instead of deleting and generating. e.g. you change a track's
filename -> you keep the original conversion and the new one. This
feature is intended for faster simple library additions: large changes
to metadata and such should always warrant a full delete+generate.